move everything to server and models
