+++ server/models/index.js
@@ -1,0 +1,11 @@
+const mongoose = require('mongoose');
+
+const PriceHistory = mongoose.model('PriceHistory', {
+    dt: Number,
+    simadi: Number,
+    date: Number
+});
+
+module.exports = {
+    PriceHistory
+}
